requirements.txt
fastapi
uvicorn
pandas
scikit-learn
pymysql
python-dotenv